Move GetRequiredDeviceSize to the device tree module
The Storage module should provide DBus API only for managing other
storage modules. The storage-related utilities should be in the
device tree module.
