+++ tests/nosetests/pyanaconda_tests/module_storage_test.py
@@ -24,7 +24,6 @@ from unittest.mock import patch, call, Mock
 
 from blivet.devicelibs.crypto import MIN_CREATE_ENTROPY
 from blivet.formats.luks import LUKS2PBKDFArgs
-from blivet.size import Size
 
 from pykickstart.constants import AUTOPART_TYPE_LVM_THINP, AUTOPART_TYPE_PLAIN, AUTOPART_TYPE_LVM
 
@@ -96,11 +95,6 @@ class StorageInterfaceTestCase(unittest.TestCase):
 
         obj.implementation.stopped_signal.emit()
         storage_changed_callback.assert_called_once()
-
-    def get_required_device_size_test(self):
-        """Test GetRequiredDeviceSize."""
-        required_size = self.storage_interface.GetRequiredDeviceSize(Size("1 GiB").get_bytes())
-        self.assertEqual(Size("1280 MiB").get_bytes(), required_size, Size(required_size))
 
     @patch('pyanaconda.modules.storage.partitioning.validate.storage_checker')
     def apply_partitioning_test(self, storage_checker):
